+++ HISTORY.rst
@@ -8,6 +8,14 @@ New Features
 ^^^^^^^^^^^^
 * ``clisops.ops.average.average_shape`` added (#312). Exposing average_shape from clisops.core to clisops.ops.
 
+Bug Fixes
+^^^^^^^^^
+* Now also applying fix for datasets with shifted longitude frames (#218) for the regrid operator (#313).
+
+Other Changes
+^^^^^^^^^^^^^
+* Warnings are now emitted when the user attempts to regrid a zonal mean dataset (#313).
+
 v0.12.1 (2023-11-30)
 --------------------